% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab03.r
\name{euclidean}
\alias{euclidean}
\title{Function to find the greatest common divisor of two numbers.}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{Integer}

\item{y}{Integer}
}
\value{
gcd of x ad y
}
\description{
This function use Euclidian algorithm to caculate a GCD of 2 numbers(x and y)
}
