% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab03-package.R
\docType{package}
\name{lab03-package}
\alias{lab03-package}
\alias{lab03}
\title{Lab 3 of Advanced Programming in R}
\description{
This package include following functions.\cr 
1. euclidean\cr 
2. dijkstra\cr 
\cr
The package also has the following data sources included.\cr 
1. wiki_graph \cr
}
