% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab03.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph:  data frame for dijkstra function}
\format{
A data frame contain the edges between nodes and distance of those edges.
\describe{
  \item{v1}{start node}
  \item{v2}{end node}
  \item{w}{edge length}
}
}
\usage{
wiki_graph
}
\description{
wiki_graph:  data frame for dijkstra function
}
\keyword{datasets}
