% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab03.r
\name{dijkstra}
\alias{dijkstra}
\title{Function to calculate the shortest path from the initial node to other node 
in the graph}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{An data frame store the edge information}

\item{init_node}{Init node number}
}
\value{
An vector that stores the distance
}
\description{
This function implements the Dijkstras algorithm and finds the shortest distance 
from the initial node to every other node in the graph
Author: Qinyuan QI
}
