% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab03.r
\name{euclidean}
\alias{euclidean}
\title{Function to find the greatest common divisor of two numbers.
This function uses Euclidian algorithm to calculate a GCD of 2 numbers(x and y)}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{Integer}

\item{y}{Integer}
}
\value{
gcd of x ad y
}
\description{
Author: Satya Sai Naga Jaya Koushik Pilla
}
